changed to builders

+++ athena/src/main/java/project/dao/athena/SetupStruct.java
@@ -3,13 +3,15 @@ package project.dao.athena;
 import project.elgamal.ElGamalPK;
 import project.elgamal.ElGamalSK;
 
+import java.math.BigInteger;
+
 public class SetupStruct {
     public final PK_Vector pkv;
     public final ElGamalSK sk;
     public final int mb;
-    public final int mc;
+    public final BigInteger mc;
 
-    public SetupStruct(PK_Vector pkv, ElGamalSK sk, int mb, int mc) {
+    public SetupStruct(PK_Vector pkv, ElGamalSK sk, int mb, BigInteger mc) {
 
         this.pkv = pkv;
         this.sk = sk;
